Add basic telemetry with tracing crate
Fixed some of the existing GitHub workflows by updating their actions.
The CI/CD pipeline needs more work and is a WIP.

The telemetry layer currently serializes data to JSON for convenience.
The `tracer` module will be moved to another repository as it continues
to grow in complexity.

Upcoming commits will clean up the repository and include refactoring to
make the project architecture and design more sensible. The project
started off as an experiment with no true intentions of building
something that may reach production. These issues will all be addressed
over the next week.